MOIA-55452: Also retry stack status fails
It makes sense to also retry when describing the stack fails. Since we
have a lot of time and we backoff exponentially this also shouldn't
contribute to overloading the API even more.

It probably also makes sense to immediately wait a little more aggressively
in order to put pressure of the CloudFormation API.